+++ src/particles/ill_bubble.cpp
@@ -28,8 +28,8 @@
 
 // Vibration period of the bubble
 const uint vib_period = 250;
-const float MAX_SCALE = 1.0f;
-const float MIN_SCALE = 0.25f;
+const Double MAX_SCALE = 1.0f;
+const Double MIN_SCALE = 0.25f;
 
 IllBubble::IllBubble() : ExplosionSmoke(20)
 {
@@ -48,10 +48,10 @@ void IllBubble::Refresh()
   ExplosionSmoke::Refresh();
   uint time = (Time::GetInstance()->Read() + vib_phi) % vib_period;
 
-  float scale_x, scale_y;
+  Double scale_x, scale_y;
   image->GetScaleFactors(scale_x, scale_y);
-  scale_x *= 1.0 + 0.2 * sin(2.0 * M_PI * time / (float)vib_period);
-  scale_y *= 1.0 + 0.2 * cos(2.0 * M_PI * time / (float)vib_period);
+  scale_x *= 1.0 + 0.2 * sin(2.0 * M_PI * time / (Double)vib_period);
+  scale_y *= 1.0 + 0.2 * cos(2.0 * M_PI * time / (Double)vib_period);
   scale_x = std::max(MIN_SCALE, std::min(scale_x, MAX_SCALE));
   scale_y = std::max(MIN_SCALE, std::min(scale_y, MAX_SCALE));
   image->Scale(scale_x, scale_y);
@@ -60,7 +60,7 @@ void IllBubble::Refresh()
 void IllBubble::Draw()
 {
   if (m_left_time_to_live > m_initial_time_to_live - 3)
-    image->SetAlpha( (float)(m_initial_time_to_live - m_left_time_to_live) / 3.0 );
+    image->SetAlpha( (Double)(m_initial_time_to_live - m_left_time_to_live) / 3.0 );
   else
     image->SetAlpha(1.0);
